Move Particle class to separate file
