+++ src/wash/wash.cpp
@@ -1,60 +1,18 @@
 #include "wash.hpp"
 
 namespace wash {
-    static uint64_t max_iterations;
-    static std::vector<std::string> forces_scalar;
-    static std::vector<std::string> forces_vector;
-    static std::vector<Kernel> init_kernels;
-    static std::vector<Kernel> loop_kernels;
-    static NeighborsFuncT neighbors_kernel;
-    static std::vector<Particle> particles;
-    static std::unordered_map<std::string, double> variables;
-    static std::string simulation_name;
-    static std::string output_file_name;
-
-    Particle::Particle(const int id) : id(id) {}
-
-    double Particle::get_density() const { return this->density; }
-
-    void Particle::set_density(const double density) { this->density = density; }
-
-    double Particle::get_mass() const { return this->mass; }
-
-    void Particle::set_mass(const double mass) { this->mass = mass; }
-
-    double Particle::get_smoothing_length() const { return this->smoothing_length; }
-
-    void Particle::set_smoothing_length(const double smoothing_length) { this->smoothing_length = smoothing_length; }
-
-    SimulationVecT Particle::get_pos() const { return this->pos; }
-
-    void Particle::set_pos(const SimulationVecT pos) { this->pos = pos; }
-
-    SimulationVecT Particle::get_vel() const { return this->vel; }
-
-    void Particle::set_vel(const SimulationVecT vel) { this->vel = vel; }
-
-    SimulationVecT Particle::get_acc() const { return this->acc; }
-
-    void Particle::set_acc(const SimulationVecT acc) { this->acc = acc; }
-
-    double Particle::get_force_scalar(const std::string& force) const { return this->force_scalars.at(force); }
-
-    void Particle::set_force_scalar(const std::string& force, const double value) {
-        this->force_scalars[force] = value;
+    namespace {
+        uint64_t max_iterations;
+        std::vector<std::string> forces_scalar;
+        std::vector<std::string> forces_vector;
+        std::vector<Kernel> init_kernels;
+        std::vector<Kernel> loop_kernels;
+        NeighborsFuncT neighbors_kernel;
+        std::vector<Particle> particles;
+        std::unordered_map<std::string, double> variables;
+        std::string simulation_name;
+        std::string output_file_name;
     }
-
-    SimulationVecT Particle::get_force_vector(const std::string& force) const { return this->force_vectors.at(force); }
-
-    void Particle::set_force_vector(const std::string& force, const SimulationVecT value) {
-        this->force_vectors[force] = value;
-    }
-
-    double Particle::get_vol() const { return get_mass() / get_density(); }
-
-    bool Particle::operator==(const Particle& other) const { return this->id == other.id; }
-
-    bool Particle::operator!=(const Particle& other) const { return !(*this == other); }
 
     ForceKernel::ForceKernel(const ForceFuncT func) : func(func) {}
